+++ utils/rfcx-guardian/guardian-msg-parsing-utils.js
@@ -186,10 +186,7 @@ exports.guardianMsgParsingUtils = {
   },
 
   assembleReceivedSegments: function (dbSegs, dbSegGrp, guardianGuid, guardianPinCode) {
-    let concatSegs = ''
-    for (let i = 0; i < dbSegs.length; i++) { concatSegs += dbSegs[i].body }
-
-    checkInHelpers.gzip.unZipJson(encodeURIComponent(concatSegs)).bind({})
+    this.decodeSegmentsToJSON(dbSegs)
       .then(function (jsonObj) {
         if (dbSegGrp.type === 'png') {
           if (hash.hashData(JSON.stringify(jsonObj)).substr(0, dbSegGrp.checksum_snippet.length) === dbSegGrp.checksum_snippet) {
@@ -218,6 +215,17 @@ exports.guardianMsgParsingUtils = {
               })
           }
         }
+      })
+  },
+
+  decodeSegmentsToJSON: function (segments) {
+    let concatSegs = ''
+    const sortedSegments = segments.sort((a, b) => a.segment_id - b.segment_id) // sort by segment_id to make segments arranged
+    for (let i = 0; i < sortedSegments.length; i++) { concatSegs += sortedSegments[i].body }
+
+    return checkInHelpers.gzip.unZipJson(encodeURIComponent(concatSegs)).bind({})
+      .then(function (jsonObj) {
+        return jsonObj
       })
   }